Simplify how indentation is configured
The existing --indent option is a mess. Giving users an option to switch
between hard and soft tabs was utterly pointless, and made a mess of the
documentation. Since this program's only purpose is to make JSON data as
readable as possible for command-line environments, there's no reason to
allow that degree of customisation.

The aliases -t and --tab have also been dropped in an effort to keep the
module as simple and focussed as possible.